[SLP][modularisation][NFC] Move BaseShuffleAnalysis to SLPShuffleAnalysis.h
Move the BoUpSLP-independent shuffle-analysis base class out of
SLPVectorizer.cpp into a new header-only SLPVectorizer/SLPShuffleAnalysis.h.
BoUpSLP::ShuffleCostEstimator and BoUpSLP::ShuffleInstructionBuilder keep
deriving from it.

createShuffle read the file-local SLPReVec cl::opt in an assert; the option
stays static in SLPVectorizer.cpp and the moved template takes its value as
an explicit bool parameter. Behavior is unchanged.

Part of the SLPVectorizer.cpp modularization effort:
https://discourse.llvm.org/t/modularizing-slpvectorizer-cpp/90922
